chore: move files to new structure
