+++ apps/client/src/views/cuesheet/ProtectedCuesheet.tsx
