+++ apps/client/src/views/cuesheet/cuesheet-progress/CuesheetProgress.tsx
