+++ apps/client/src/views/cuesheet/cuesheet-table-elements/CuesheetHeader.tsx
