+++ apps/client/src/views/cuesheet/cuesheet-table-elements/EditableCell.tsx
