+++ apps/client/src/views/cuesheet/cuesheet-table-elements/EventRow.tsx
