+++ apps/client/src/views/cuesheet/cuesheet-table-settings/CuesheetTableSettings.tsx
