+++ apps/client/src/views/cuesheet/useColumnManager.tsx
